fixed tickets reducer removed ticket reducer
