+++ .gitignore
@@ -1,2 +1,3 @@
 node_modules/
 /config/keys.js
+/frontend/node_modules/
