+++ frontend/src/reducers/ticket_errors_reducer.js
@@ -1,8 +1,7 @@
-import { RECEIVE_TICKET_ERRORS, RECEIVE_TICKET } from '../actions/ticket_actions'
+import { RECEIVE_TICKET_ERRORS } from '../actions/ticket_actions'
 
 const ticketErrorsReducer = (defaultState = {}, action) => {
     Object.freeze(defaultState)
-    const newState = Object.assign({}, defaultState)
 
     switch (action.type) {
         case RECEIVE_TICKET_ERRORS:
