--- frontend/src/reducers/ticket_reducer.js
@@ -1,18 +1,0 @@
-import { RECEIVE_TICKETS, RECEIVE_TICKET } from "../actions/ticket_actions";
-
-const ticketReducer = (defaultState = {}, action) => {
-    Object.freeze(defaultState)
-    const newState = Object.assign({}, defaultState)
-
-    switch (action.type) {
-        case RECEIVE_TICKETS:
-            return action.tickets
-        case RECEIVE_TICKET:
-            newState[action.ticket.id] = action.ticket
-            return newState
-        default:
-            return defaultState
-    }
-}
-
-export default ticketReducer;
