+++ frontend/src/reducers/tickets_reducer.js
@@ -1,10 +1,17 @@
+import { RECEIVE_TICKETS, RECEIVE_TICKET } from "../actions/ticket_actions";
 
-const ticketsReducer = (state = {}, action) => {
+const ticketsReducer = (defaultState = {}, action) => {
+    Object.freeze(defaultState)
+    const newState = Object.assign({}, defaultState)
+
     switch (action.type) {
-        case "banana":
-            return state;
+        case RECEIVE_TICKETS:
+            return action.tickets
+        case RECEIVE_TICKET:
+            newState[action.ticket.id] = action.ticket
+            return newState
         default:
-            return state;
+            return defaultState
     }
 }
